successfully combined add/edit pup into PupForm
